reorganize grib definitions handling
