+++ src/epygram/extra/griberies/definitions/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) Météo France (2014-)
+# This software is governed by the CeCILL-C license under French law.
+# http://www.cecill.info
+"""
+Contains helpers to handle GRIB definitions.
+"""
+
+from .names import NamesGribDef
+
+namesgribdef = NamesGribDef(actual_init=False)
+
